docs: move directions of step-by-step guide to exmaple code to reduce redundant
