+++ docs/get-started.md
@@ -56,117 +56,9 @@ External locks are used to receive CKB or tokens from other users. Internal lock
 
 Live Cells are the cells that are alive on the CKB chain, they can be spent by the controller of its lock. It's just another name for Unspent-Transaction-Output (UTXO).
 
-## Issue a transfer transaction with the Nexus Wallet
+## The First Example
 
-We recommend you to read the step-by-step guide in the following section to get a better understanding of how to use the Nexus Wallet. But if you would like to jump into the code directly, you can check out the [online example](https://githubbox.com/ckb-js/nexus/tree/develop/examples/nexus-example).
-
-### Step 1: Enable the Nexus Wallet
-
-Firstly, you need to enable the Nexus Wallet by calling the `wallet_enable` RPC.
-
-If your dApp is the first time to connect to the wallet,
-you'll need to call `wallet_enable` API to enable the wallet for the dApp.
-After calling this method, we can see the dApp in the whitelist
-
-```js
-await window.ckb.request({ method: 'wallet_enable' });
-```
-
-### Step 2: Get one off-chain lock and receive some testnet CKB
-
-Then you can get the off-chain locks by calling the `wallet_fullOwnership_getOffChainLocks` RPC.
-
-```js
-const offChainLocks = await window.ckb.request({
-  method: 'wallet_fullOwnership_getOffChainLocks',
-  params: { change: 'external' },
-});
-const lock = offChainLocks[0];
-const address = helpers.encodeToAddress(lock);
-```
-
-Then you can use the address to receive some testnet CKB from the [CKB Faucet](https://faucet.nervos.org/).
-
-After you have received the testnet CKB, the lock that used to receive the CKB will become an on-chain lock.
-
-### Step 3: Get the live cells and build a transfer transaction
-
-Now that you 10000 CKB, let's transfer 100 CKB to another address.
-
-We can get the live cells by calling the `wallet_fullOwnership_getLiveCells` RPC.
-
-```typescript
-// use an external lock to receive the CKB
-const externalOffChainLocks = await window.ckb.request({
-  method: 'wallet_fullOwnership_getOffChainLocks',
-  params: { change: 'external' },
-});
-const receiveLock = externalOffChainLocks[0];
-// use an external lock to receive the CKB change
-const internalOffChainLocks = await window.ckb.request({
-  method: 'wallet_fullOwnership_getOffChainLocks',
-  params: { change: 'internal' },
-});
-const changeLock = internalOffChainLocks[0];
-// get the live cells to spend
-const getLiveCellsRes = await window.ckb.request({
-  method: 'wallet_fullOwnership_getLiveCells',
-  params: {},
-});
-const liveCells: Cell[] = getLiveCellsRes.objects;
-const totalCapacity = liveCells.reduce((acc, cell) => acc.add(cell.cellOutput.capacity), BI.from(0));
-let txSkeleton = helpers.TransactionSkeleton();
-// setup inputs
-txSkeleton = txSkeleton.update('inputs', (inputs) => {
-  return inputs.concat(...liveCells);
-});
-// setup outputs
-txSkeleton = txSkeleton.update('outputs', (outputs) => {
-  return outputs.concat(
-    // transfer 100 CKB to the receivelock
-    {
-      cellOutput: {
-        capacity: BI.from(100)
-          .mul(10 ** 8)
-          .toHexString(),
-        lock: receiveLock,
-      },
-      data: '0x',
-    },
-    // receive change CKB
-    {
-      cellOutput: {
-        capacity: totalCapacity
-          .sub(BI.from(100).mul(10 ** 8))
-          .sub(1000) // 1000 shannons for tx fee
-          .toHexString(),
-        lock: changeLock,
-      },
-      data: '0x',
-    },
-  );
-});
-
-// to build the transaction, there are still some other things to do
-// fill in cellDeps, placeholder witnesses, etc.
-```
-
-### Step 4: Sign the transaction and send it to the CKB chain
-
-After you have built the transactionSkeleton, you can sign the transaction by calling the `wallet_fullOwnership_signTransaction` RPC.
-
-```typescript
-// sign the transaction with Nexus Wallet RPC
-const tx = helpers.createTransactionFromSkeleton(txSkeleton);
-const signatures: [Script, HexString] = await ckb.fullOwnership.signTransaction({ tx });
-// put signatures into witnesses
-for (let index = 0; index < signatures.length; index++) {
-  const [_, sig] = signatures[index];
-  tx.witnesses[index] = bytes.hexify(blockchain.WitnessArgs.pack({ lock: sig }));
-}
-// send the transaction
-const txHash = await rpc.sendTransaction(tx);
-```
+You can check out the [online example](https://githubbox.com/ckb-js/nexus/tree/develop/examples/nexus-example) to have a hands-on experience of Nexus Wallet.
 
 ## Other references
